Revisi

+++ models/product_models.go
@@ -7,8 +7,8 @@ import (
 type Product struct {
 	gorm.Model
 	Title      string   `json:"title"`
-	Price      int      `json:"price"`
-	Stock      int      `json:"stock"`
+	Price      float64  `json:"price"`
+	Stock      float64  `json:"stock"`
 	CategoryID int      `json:"category_id"`
 	Category   Category `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 }
